update qwenvl tests
